Retrieve service catalog when it is missing
The Heat API uses the keystone endpoint GET /{version}/tokens/{token id}
to validate tokens. The specification for the v2.0 version of this
endpoint omits the service catalog in its response (although the
OpenStack implementation of Keystone v2 still includes it). The
response is passed along to the Heat engine which uses it as part of its
RequestContext. Heat expects the service catalog to be in this context,
so this patch retrieves it if it is not there.

closes-bug: #1420987

Change-Id: I183b22931fa8c997883168038e260db3d54b5fe5