Remove unconditional return of user_id
This fixes issues with subusers using software config. Heat should not
attempt to delete the stack user if one had not been created before.
The code that this patch removes caused an ID to always be returned.
Additionally, it set the ID in the resource metadata whether a stack
user existed or not, making it impossible to tell if one existed only
from data within the resource.

Change-Id: I50804998fdce1f7de06906ae9a2f47317721a6f6
Closes-bug: #1531606